refactor: optimized code

+++ apps/api/src/utils/playgroundUtils.ts
@@ -5,6 +5,7 @@ import fs from "fs-extra"
 import util from "util"
 import { exec } from 'child_process';
 import { redis } from "./redis";
+import { checkTag } from "./containerUtils";
 
 const execAsync = util.promisify(exec);
 
@@ -53,17 +54,33 @@ export const setupPlayground = async (framework: FrameworkType, tag: string) => 
 };
 
 export const clearPlayground = async (tag: string) => {
-    try {
-        await execAsync(`docker stop ${tag}`);
+    const containerExists = await checkTag(tag);
+    if (containerExists) {
+        try {
+            await execAsync(`docker stop ${tag}`);
+            console.log("Container Stopped")
+            await execAsync(`docker rm ${tag}`);
+            console.log("Container Removed")
+            await execAsync(`rm -rf ./tmp/${tag}`)
+            console.log("Files Removed")
+            await redis.del(tag);
+        }
+        catch (e) {
+            console.error(e);
+        }
+
     }
-    catch (e) {
-        console.error(e);
+}
+const deleteFolderRecursive = async (path: string) => {
+    if (fs.existsSync(path)) {
+        fs.readdirSync(path).forEach((file, index) => {
+            const curPath = `${path}/${file}`;
+            if (fs.lstatSync(curPath).isDirectory()) { // recurse
+                deleteFolderRecursive(curPath);
+            } else { // delete file
+                fs.unlinkSync(curPath);
+            }
+        });
+        fs.rmdirSync(path);
     }
-    try {
-        await execAsync(`docker rm ${tag}`);
-    }
-    catch (e) {
-        console.error(e);
-    }
-    await redis.del(tag);
 }
